* commit '29eadb328a71f912b0d479860286cd606ed3dc87':
  AArch64: Use long for pointers in hardware classes

+++ core/java/android/hardware/Camera.java
@@ -150,7 +150,7 @@ public class Camera {
     private static final int CAMERA_MSG_PREVIEW_METADATA = 0x400;
     private static final int CAMERA_MSG_FOCUS_MOVE       = 0x800;
 
-    private int mNativeContext; // accessed by native methods
+    private long mNativeContext; // accessed by native methods
     private EventHandler mEventHandler;
     private ShutterCallback mShutterCallback;
     private PictureCallback mRawImageCallback;
